+++ Packages/UGF.Code.Generate/Editor/Container.External/CodeGenerateContainerExternalInfo.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace UGF.Code.Generate.Editor.Container.External
+{
+    /// <summary>
+    /// Represents default implementation of the container external type information.
+    /// </summary>
+    [Serializable]
+    public class CodeGenerateContainerExternalInfo : CodeGenerateContainerExternalInfoBase<CodeGenerateContainerExternalMemberInfo>
+    {
+    }
+}